- Create tests/ module - Set example DAG to use jinja templates - Modify Fivetran sensor to not call live method during init
