+++ fivetran_provider/example_dags/example_fivetran.py
@@ -18,14 +18,14 @@ dag = DAG(
 fivetran_sync_start = FivetranOperator(
     task_id='fivetran-task',
     fivetran_conn_id='fivetran',
-    connector_id=Variable.get('connector_id'),
+    connector_id="{{ var.value.get('connector_id', 'fallback_connector_id') }}",
     dag=dag
 )
 
 fivetran_sync_wait = FivetranSensor(
     task_id='fivetran-sensor',
     fivetran_conn_id='fivetran',
-    connector_id=Variable.get('connector_id'),
+    connector_id="{{ var.value.get('connector_id', 'fallback_connector_id') }}",
     poke_interval=5,
     dag=dag
 )
