+++ fivetran_provider/sensors/fivetran.py
@@ -34,7 +34,7 @@ class FivetranSensor(BaseSensorOperator):
         self.connector_id = connector_id
         self.poke_interval = poke_interval
         self.hook = FivetranHook(self.fivetran_conn_id)
-        self.previous_completed_at = self.hook.get_last_sync(self.connector_id)
 
     def poke(self, context):
-       return self.hook.get_sync_status(self.connector_id, self.previous_completed_at)
+        self.previous_completed_at = self.hook.get_last_sync(self.connector_id)
+        return self.hook.get_sync_status(self.connector_id, self.previous_completed_at)
